# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aryan Landmark Solutions API
  version: 1.0.0
  description: |
    Production REST API for Aryan Landmark Solutions.

    PostgreSQL contains six tables. All property photos, property documents
    and inquiry call recordings use one unified `attachments` resource.

    Cloudflare R2 stores file binaries. PostgreSQL stores metadata and the R2
    object key.
servers:
  - url: /api/v1
...
paths:
...
  /master/users:
    get:
      tags: [Master Management]
      summary: List ADMIN/EMPLOYEE users (Master Admin user management)
      description: |
        Phase 15A. Same query semantics as GET /users (page, pageSize,
        role, status, search — all optional) reachable under the
        master-JWT boundary, so a MASTER account (which cannot hold an
        application JWT) can still list users for the Master Admin UI.
      security:
        - masterBearer: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: role
          in: query
          schema:
            type: string
            enum: [ADMIN, EMPLOYEE]
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, BLOCKED]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Master Management]
      summary: Create ADMIN or EMPLOYEE
      security:
        - masterBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /users:
    get:
      tags: [Users]
      summary: List application users
      security:
        - applicationBearer: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: role
          in: query
          schema:
            type: string
            enum: [ADMIN, EMPLOYEE]
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, BLOCKED]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
  /users/{id}:
    patch:
      tags: [Users]
      summary: Change an employee's status (activate/deactivate/block)
      description: |
        Phase 35. ADMIN-only — an EMPLOYEE token gets 403 here, unlike
        GET /users above which both roles may call. Scoped to
        userType=APPLICATION_USER: a CUSTOMER or MASTER id, or an
        unknown id, both resolve to 404 USER_NOT_FOUND (never leaking
        which of those it was). The request body accepts `status` only
        — any other field (`role`, `email`, `password`, etc.) is
        rejected with 400, not silently ignored. This is a deliberately
        narrower sibling of PATCH /master/users/{id} (MASTER-only, can
        change far more fields) — ADMIN gets exactly the one capability
        this phase approved and nothing else.

        Takes effect immediately, not just on the target's next login:
        requireApplicationAuth re-checks `users.status` on every
        authenticated request, so a status other than ACTIVE ends that
        user's current session as soon as their device next reaches the
        backend (see LoginResponse's `expiresIn` description). A fully
        offline device cannot know about this until connectivity
        returns — this is expected, not a bug, and no maximum offline
        duration is enforced.
      security:
        - applicationBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: User status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT

    masterBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authorization failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Business conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateApplicationUserRequest:
      type: object
      required: [name, mobile, password, role]
      properties:
        userId:
          type: string
          example: EMP001
        name:
          type: string
        email:
          type: string
          format: email
        mobile:
          type: string
        alternateMobile:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        password:
          type: string
          format: password
          minLength: 8
        role:
          type: string
          enum: [ADMIN, EMPLOYEE]
...
    User:
      type: object
      required: [id, userType, name, status]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          nullable: true
        userType:
          type: string
          enum: [APPLICATION_USER, CUSTOMER, MASTER]
        role:
          type: string
          nullable: true
          enum: [ADMIN, EMPLOYEE]
        name:
          type: string
        email:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BLOCKED]
...
    UpdateUserStatusRequest:
      type: object
      required: [status]
      description: |
        Phase 35 — the request body for PATCH /users/{id}. `status` is
        the only accepted field; the endpoint returns 400 if any other
        key (role, email, password, etc.) is present in the body at all,
        rather than silently ignoring it.
      properties:
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BLOCKED]
      additionalProperties: false
    UserResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    UserListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
        requestId:
          type: string
...